Fixed all: warning C4244: 'argument' : conversion from '...' to '...', possible loss of data errors
Voornamelijk time_t problemen
- o.a. alle srand(time(NULL)) omgezet naar: srand(static_cast<unsigned
int>(time(NULL)));
- Warning van DWORD in game.cpp regel 88 is ook opgelost -> @PimV  keer
naar laten kijken.

+++ Bermuda/Bermuda/Game.cpp
@@ -85,7 +85,7 @@ void Game::gameLoop(GameStateManager* gsm) {
 		QueryPerformanceCounter(&afterLoopTime);
 		//Sleep if needed (if time took longer than optimal)
 		if ((((previousTime.QuadPart - afterLoopTime.QuadPart) * 1000.0 / frequency.QuadPart) + OPTIMAL_TIME) > 0) {
-			Sleep((((previousTime.QuadPart - afterLoopTime.QuadPart) * 1000.0 / frequency.QuadPart) + OPTIMAL_TIME));
+			Sleep(static_cast<DWORD>((((previousTime.QuadPart - afterLoopTime.QuadPart) * 1000.0 / frequency.QuadPart) + OPTIMAL_TIME)));
 		}
 	}
 }
